Fix tg_bd

+++ tg_bot/src/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Boolean, Column, Integer
+from sqlalchemy import Boolean, Column, BigInteger
 from sqlalchemy.ext.declarative import declarative_base
 
 Base = declarative_base()
@@ -6,6 +6,5 @@ Base = declarative_base()
 
 class User(Base):
     __tablename__ = 'users'
-
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
     status = Column(Boolean, default=False, nullable=False)
